+++ release-notes.md
@@ -2,6 +2,8 @@
 * NEW FEATURE: Add ability to show minion group (max) size
 * NEW FEATURE: Add ability to use actor dice pool with vehicle weapon rolls
 * Fix for shop item finding pool being built incorrectly
+* Fix for non-working adversary status to apply
+* Improve attack animation: players no longer need to select their own token to play attack animations (if they only have one token on the canvas)
 
 `0.1.11` - 2021-06-29
 * Fix for missing localizations ([#50](https://github.com/wrycu/StarWarsFFG-Enhancements/issues/50))